Update to accommodate changed maidsafe-dht API.
git-svn-id: svn+ssh://dash.maidsafe.net/srv/svn/dawn/trunk@3484 6165b5c9-d236-0410-b717-85aa93a55b26

+++ src/tests/maidsafe/testprivateshares.cc
@@ -386,13 +386,14 @@ TEST_F(PrivateSharesTest, BEH_MAID_MI_Create_ListShares) {
   // Test share list to be empty
   std::list<maidsafe::private_share> share_list;
   ASSERT_EQ(0, psh_->MI_GetShareList(&share_list)) << "Failed getting list.";
-  ASSERT_EQ(0, share_list.size()) << "Share container not empty on creation.";
+  ASSERT_EQ(static_cast<unsigned int>(0), share_list.size()) <<
+            "Share container not empty on creation.";
 
   // Test full share list to be empty
   std::list<maidsafe::PrivateShare> full_share_list;
   ASSERT_EQ(0, psh_->MI_GetFullShareList(&full_share_list)) <<
             "Failed getting full list";
-  ASSERT_EQ(0, full_share_list.size()) <<
+  ASSERT_EQ(static_cast<unsigned int>(0), full_share_list.size()) <<
             "Share container not empty on creation.";
 
   // Test lower bound of field index
@@ -403,7 +404,7 @@ TEST_F(PrivateSharesTest, BEH_MAID_MI_Create_ListShares) {
   ASSERT_EQ("", ps.Msid()) << "Share container Msid was modified.";
   ASSERT_EQ("", ps.MsidPubKey()) << "Share container MsidPubKey was modified.";
   ASSERT_EQ("", ps.MsidPriKey()) << "Share container MsidPriKey was modified.";
-  ASSERT_EQ(0, ps.Participants().size()) <<
+  ASSERT_EQ(static_cast<unsigned int>(0), ps.Participants().size()) <<
             "Share container Participants was modified.";
 
   // Test upper bound of field index
@@ -413,7 +414,7 @@ TEST_F(PrivateSharesTest, BEH_MAID_MI_Create_ListShares) {
   ASSERT_EQ("", ps.Msid()) << "Share container Msid was modified.";
   ASSERT_EQ("", ps.MsidPubKey()) << "Share container MsidPubKey was modified.";
   ASSERT_EQ("", ps.MsidPriKey()) << "Share container MsidPriKey was modified.";
-  ASSERT_EQ(0, ps.Participants().size()) <<
+  ASSERT_EQ(static_cast<unsigned int>(0), ps.Participants().size()) <<
             "Share container Participants was modified.";
 
   // Test wrong share name
@@ -423,7 +424,7 @@ TEST_F(PrivateSharesTest, BEH_MAID_MI_Create_ListShares) {
   ASSERT_EQ("", ps.Msid()) << "Share container Msid was modified.";
   ASSERT_EQ("", ps.MsidPubKey()) << "Share container MsidPubKey was modified.";
   ASSERT_EQ("", ps.MsidPriKey()) << "Share container MsidPriKey was modified.";
-  ASSERT_EQ(0, ps.Participants().size()) <<
+  ASSERT_EQ(static_cast<unsigned int>(0), ps.Participants().size()) <<
             "Share container Participants was modified.";
 
   // Test wrong share msid
@@ -433,28 +434,31 @@ TEST_F(PrivateSharesTest, BEH_MAID_MI_Create_ListShares) {
   ASSERT_EQ("", ps.Msid()) << "Share container Msid was modified.";
   ASSERT_EQ("", ps.MsidPubKey()) << "Share container MsidPubKey was modified.";
   ASSERT_EQ("", ps.MsidPriKey()) << "Share container MsidPriKey was modified.";
-  ASSERT_EQ(0, ps.Participants().size()) <<
+  ASSERT_EQ(static_cast<unsigned int>(0), ps.Participants().size()) <<
             "Share container Participants was modified.";
 
   // Test wrong index for field in share participant lookup
   std::list<maidsafe::share_participant> sp_list;
   ASSERT_EQ(-2015, psh_->MI_GetParticipantsList("aaa", -1, &sp_list)) <<
             "Failed to recognise lower bound.";
-  ASSERT_EQ(0, sp_list.size()) << "List should have remained empty.";
+  ASSERT_EQ(static_cast<unsigned int>(0), sp_list.size()) <<
+            "List should have remained empty.";
   ASSERT_EQ(-2015, psh_->MI_GetParticipantsList("aaa", 2, &sp_list)) <<
             "Failed to recognise upper bound.";
-  ASSERT_EQ(0, sp_list.size()) << "List should have remained empty.";
+  ASSERT_EQ(static_cast<unsigned int>(0), sp_list.size()) <<
+            "List should have remained empty.";
 
   // Test wrong share name for participant list
   ASSERT_EQ(-2015, psh_->MI_GetParticipantsList("aaa", 0, &sp_list)) <<
             "Failed to recognise invalid share name.";
-  ASSERT_EQ(0, sp_list.size()) << "List should have remained empty.";
+  ASSERT_EQ(static_cast<unsigned int>(0), sp_list.size()) <<
+            "List should have remained empty.";
 
   // Test wrong share msid for participant list
   ASSERT_EQ(-2015, psh_->MI_GetParticipantsList("aaa", 1, &sp_list)) <<
             "Failed to recognise invalid share msid.";
-  ASSERT_EQ(0, sp_list.size()) << "List should have remained empty.";
-
+  ASSERT_EQ(static_cast<unsigned int>(0), sp_list.size()) <<
+            "List should have remained empty.";
 }
 
 /*
